+++ Magic/src/main/java/com/elmakers/mine/bukkit/automata/Caster.java
@@ -1,0 +1,44 @@
+package com.elmakers.mine.bukkit.automata;
+
+import java.util.ArrayDeque;
+import java.util.Arrays;
+import java.util.Deque;
+
+import javax.annotation.Nonnull;
+
+import org.apache.commons.lang.StringUtils;
+import org.bukkit.configuration.ConfigurationSection;
+
+import com.elmakers.mine.bukkit.api.spell.Spell;
+import com.elmakers.mine.bukkit.magic.Mage;
+import com.elmakers.mine.bukkit.utility.RandomUtils;
+import com.elmakers.mine.bukkit.utility.WeightedPair;
+
+public class Caster {
+    @Nonnull
+    private final Deque<WeightedPair<String>> spells;
+
+    public Caster(@Nonnull AutomatonTemplate automaton, ConfigurationSection configuration) {
+        spells = new ArrayDeque<>();
+        RandomUtils.populateStringProbabilityMap(spells, configuration, "spells");
+    }
+
+    public void cast(Mage mage) {
+        String castSpell = RandomUtils.weightedRandom(spells);
+        if (castSpell.length() > 0) {
+            String[] parameters = null;
+            Spell spell = null;
+            if (!castSpell.equalsIgnoreCase("none")) {
+                if (castSpell.contains(" ")) {
+                    parameters = StringUtils.split(castSpell, ' ');
+                    castSpell = parameters[0];
+                    parameters = Arrays.copyOfRange(parameters, 1, parameters.length);
+                }
+                spell = mage.getSpell(castSpell);
+            }
+            if (spell != null) {
+                spell.cast(parameters);
+            }
+        }
+    }
+}